feat: implement WSStmt2 request and response classes
