+++ src/Driver/Impl/StmtBuilder/VariableLengthBuilder.cs
@@ -8,32 +8,15 @@ namespace TDengine.Driver.Impl.StmtBuilder
     {
         public TDengineDataType DataType { get; }
         public int Length { get; private set; }
-        public bool IsVariable => true;
-        public List<int> LengthList { get; } = new List<int>(1);
+        private List<int> LengthList { get; } = new List<int>(1);
         private int _totalValueLength = 0;
+        private List<byte> Values { get; set; } = new List<byte>(0);
+        private List<byte> NullMem { get; set; } = new List<byte>();
+        private int NullCount { get; set; }
 
         public VariableLengthBuilder(TDengineDataType dataType)
         {
             DataType = dataType;
-        }
-
-        private List<byte> Values { get; set; } = new List<byte>(0);
-        private List<byte> NullMem { get; set; } = new List<byte>();
-        public int NullCount { get; private set; }
-
-        public void AppendObject(object value)
-        {
-            switch (value)
-            {
-                case string s:
-                    AppendString(s);
-                    break;
-                case byte[] b:
-                    AppendBytes(b);
-                    break;
-                default:
-                    throw new ArgumentException($"Unsupported type: {value.GetType()}");
-            }
         }
 
         public void AppendString(string value)
@@ -51,7 +34,8 @@ namespace TDengine.Driver.Impl.StmtBuilder
             {
                 NullMem.Add(0);
             }
-            Length+=1;
+
+            Length += 1;
         }
 
         public void AppendNull()
@@ -60,20 +44,11 @@ namespace TDengine.Driver.Impl.StmtBuilder
             {
                 NullMem = new List<byte>(new byte[Length]);
             }
+
             LengthList.Add(0);
             NullMem.Add(1);
             NullCount += 1;
-            Length+=1;
-        }
-
-        public int ValueLength()
-        {
-            return _totalValueLength;
-        }
-
-        public int NullLength()
-        {
-            return Length;
+            Length += 1;
         }
 
         public void Clear()
@@ -83,8 +58,9 @@ namespace TDengine.Driver.Impl.StmtBuilder
             NullMem.Clear();
             NullCount = 0;
             _totalValueLength = 0;
+            Length = 0;
         }
-        
+
         private byte[] ValueBuffer()
         {
             return Values.ToArray();
@@ -109,7 +85,7 @@ namespace TDengine.Driver.Impl.StmtBuilder
                                4 + // Num field length
                                (uint)(Length) + // IsNull field length
                                1 + // HaveLength field length
-                               (uint)(Length * 4)+ // Length field length, each length is 4 bytes
+                               (uint)(Length * 4) + // Length field length, each length is 4 bytes
                                4 + // BufferLength field length
                                (uint)valueBuffer.Length; // Buffer field length
             return new Stmt2BindColInfo
@@ -124,7 +100,7 @@ namespace TDengine.Driver.Impl.StmtBuilder
                 Buffer = valueBuffer
             };
         }
-        
+
         public Stmt2BindColInfo AddToStmt2BindColInfo(Stmt2BindColInfo bindColInfo)
         {
             if (bindColInfo.DataType != (int)DataType)
@@ -149,17 +125,62 @@ namespace TDengine.Driver.Impl.StmtBuilder
             Array.Resize(ref bindColInfo.Buffer, (int)newBufferLength);
             Buffer.BlockCopy(valueBuffer, 0, bindColInfo.Buffer, bindColInfo.Buffer.Length, valueBuffer.Length);
             bindColInfo.BufferLength += newBufferLength;
-            
+
             // Length
             var dataLength = DataLength();
             Array.Resize(ref bindColInfo.Length, bindColInfo.Num);
-            Buffer.BlockCopy(dataLength,0,bindColInfo.Length, bindColInfo.Num * 4, Length * 4);
+            Buffer.BlockCopy(dataLength, 0, bindColInfo.Length, bindColInfo.Num * 4, Length * 4);
 
             // TotalLength
             bindColInfo.TotalLength += (uint)Length // length of IsNull
                                        + (uint)valueBuffer.Length + // length of Buffer
                                        +(uint)(Length * 4); // length of Length
             return bindColInfo;
+        }
+
+        public void Remove(int count)
+        {
+            if (count < 0 || count > Length)
+            {
+                throw new ArgumentOutOfRangeException(nameof(count),
+                    "Count must be non-negative and less than or equal to the number of elements in Mem.");
+            }
+
+            var removeByteCount = 0;
+            var removeIndex = Length - count;
+            for (int i = removeIndex; i < Length; i++)
+            {
+                removeByteCount += LengthList[i];
+            }
+
+            if (removeByteCount > 0)
+            {
+                Values.RemoveRange(Length - removeByteCount, removeByteCount);
+                _totalValueLength -= removeByteCount;
+            }
+            if (NullCount > 0)
+            {
+                for (int i = removeIndex; i < Length; i++)
+                {
+                    if (NullMem[i] == 1)
+                    {
+                        NullCount--;
+                    }
+                }
+
+                if (NullCount == 0)
+                {
+                    NullMem = null;
+                }
+                else
+                {
+                    NullMem.RemoveRange(removeIndex, count);
+                }
+            }
+
+            NullCount -= count;
+            Length -= count;
+            LengthList.RemoveRange(removeIndex,count);
         }
     }
 }
