+++ src/Driver/Impl/WebSocketMethods/Protocol/WSStmt2UseResultReq.cs
@@ -1,0 +1,11 @@
+﻿using Newtonsoft.Json;
+
+namespace TDengine.Driver.Impl.WebSocketMethods.Protocol
+{
+    public class WSStmt2UseResultReq
+    {
+        [JsonProperty("req_id")] public ulong ReqId { get; set; }
+
+        [JsonProperty("stmt_id")] public ulong StmtId { get; set; }
+    }
+}
