--- src/Driver/Impl/WebSocketMethods/Protocol/WSStmtAddBatchReq.cs
@@ -1,13 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace TDengine.Driver.Impl.WebSocketMethods.Protocol
-{
-    public class WSStmtAddBatchReq
-    {
-        [JsonProperty("req_id")]
-        public ulong ReqId { get; set; }
-
-        [JsonProperty("stmt_id")]
-        public ulong StmtId { get; set; }
-    }
-}
